Tema 5.12

+++ app-contact-manager/modules/stage.js
@@ -3,14 +3,19 @@ import {
   addContact,
   addPet,
   deleteContact,
+  deletePet,
   editContact,
+  editPet,
   getContact,
+  getPet,
 } from './query.js';
 import renderMessage from './message.js';
 import { render as renderEditContact } from './editContactForm.js';
 import { render as renderAppForm } from './addPetForm.js';
+import { render as renderEditPetForm } from './editPetForm.js';
 
 const stage = document.querySelector('.stage');
+let currentContactId;
 
 // reset stage on logo click
 // aici trebuia sa scoatem si notificarea?
@@ -30,13 +35,21 @@ stage.addEventListener('click', (event) => {
     return;
   }
 
-  const button = target;
-  const parent = button.parentElement;
-  const contactId = parent.dataset.contactId;
+  // ask user to confirm deletion
+  const confirmDelete = window.confirm(
+    'Are you sure you want to delete this contact?',
+  );
+  if (confirmDelete) {
+    const button = target;
+    const parent = button.parentElement;
+    const contactId = parent.dataset.contactId;
 
-  deleteContact(contactId);
-  parent.remove();
-  addMessage(renderMessage('Contact removed', 'danger'));
+    deleteContact(contactId);
+    parent.remove();
+    addMessage(renderMessage('Contact removed', 'danger'));
+  } else {
+    return;
+  }
 });
 
 // add contact form
@@ -200,6 +213,90 @@ stage.addEventListener('submit', (event) => {
       'success',
     ),
   );
+});
+
+// delete pet
+stage.addEventListener('click', (event) => {
+  const { target } = event;
+
+  if (
+    target.nodeName !== 'BUTTON' ||
+    !target.classList.contains('delete-pet-button')
+  ) {
+    return;
+  }
+  // ask user to confirm deletion
+  const confirmDelete = window.confirm(
+    'Are you sure you want to delete this pet?',
+  );
+  if (confirmDelete) {
+    const button = target;
+    const container = button.closest('.pet');
+    const petId = Number(container.dataset.petId);
+    const contactContainer = button.closest('.contact');
+    const contactId = contactContainer.dataset.contactId;
+
+    deletePet(contactId, petId);
+    addMessage(renderMessage(`Pet was removed from list.`, 'danger'));
+
+    container.remove();
+  }
+});
+
+// edit  pet
+stage.addEventListener('click', (event) => {
+  const { target } = event;
+
+  if (
+    target.nodeName !== 'BUTTON' ||
+    !target.classList.contains('edit-pet-button')
+  ) {
+    return;
+  }
+
+  const editPetButton = target;
+  const petContainer = editPetButton.closest('.pet');
+  const petId = petContainer.dataset.petId;
+  const ownerContainer = editPetButton.closest('.contact');
+  const contactId = ownerContainer.dataset.contactId;
+  currentContactId = contactId;
+
+  const pet = getPet(contactId, petId);
+
+  if (pet === undefined) {
+    return;
+  }
+
+  clearMessages();
+  stage.innerHTML = '';
+  stage.append(renderEditPetForm(pet));
+});
+
+// save edit pet form
+stage.addEventListener('submit', (event) => {
+  const { target } = event;
+
+  if (target.nodeName !== 'FORM' || !target.classList.contains('edit-pet')) {
+    return;
+  }
+
+  event.preventDefault();
+
+  const form = target;
+
+  const { name: petName, age, species, id } = form;
+
+  const pet = {
+    name: petName.value,
+    age: age.value,
+    species: species.value,
+    id: Number(id.value),
+  };
+
+  editPet(currentContactId, pet);
+
+  stage.innerHTML = '';
+  addMessage(renderMessage(`Pet ${petName.value} saved.`, 'success'));
 });
 
 export default stage;
